Decompose VitalsManager to conform to SOLID principles
Decode dry weight
